tcq_agents.py
import os
from openai import OpenAI
from dotenv import load_dotenv
import numpy as np
import time

# Load environment variables
load_dotenv()

# Set up OpenAI client
client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

class TCQAG1:
    """
    Primary TCQ Agent (TCQAG-1) that delegates tasks and explains the TCQ process
    """
    def __init__(self):
        self.name = "TCQAG-1"
        self.role = "Task-Circuit Quantization Delegator"

    def explain_tcq(self, bit_width, outlier_percentage):
        """
        Provides an explanation of the TCQ process
        """
        prompt = f"""
        You are TCQAG-1, an AI agent specialized in explaining Task-Circuit Quantization (TCQ).

        Provide a clear, concise explanation of TCQ with the following details:
        - What TCQ is and how it differs from standard quantization
        - How TCQ identifies important weights using the saliency metric
        - The two components of the saliency metric: Quantization-aware Localization (QAL) and Magnitude-sharpened Gradient (MSG)
        - The mathematical formula: TACQ(Wij) = |Wij| · ∂L/∂Wij · |Q(Wij) - Wij|
        - How TCQ preserves {outlier_percentage}% of weights in 16-bit while quantizing the rest to {bit_width}-bit
        - Why this approach is effective for maintaining model performance
        - Real-world examples of how TCQ improves performance on specific tasks

        Keep your explanation accessible to a technical audience but avoid excessive jargon.
        Include specific numerical examples from the paper, such as performance improvements on MMLU, GSM8k, and Spider datasets.
        """

        try:
            response = client.chat.completions.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "You are TCQAG-1, an AI agent specialized in explaining Task-Circuit Quantization."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.7,
                max_tokens=1000
            )
            return response.choices[0].message.content
        except Exception as e:
            return f"Error generating explanation: {str(e)}\n\nFallback explanation: Task-Circuit Quantization (TCQ) is a novel approach that identifies and preserves critical weights in a neural network while quantizing less important ones to lower bit-widths. It uses a saliency metric combining Quantization-aware Localization and Magnitude-sharpened Gradient to determine which {outlier_percentage}% of weights to keep at 16-bit precision while quantizing the rest to {bit_width}-bit."

    def delegate_task(self, bit_width, outlier_percentage, task_type):
        """
        Delegates the TCQ demonstration task to TCQAG-2
        """
        prompt = f"""
        You are TCQAG-1, the primary AI agent for Task-Circuit Quantization demonstration.

        Create a message delegating a TCQ demonstration task to TCQAG-2 with the following parameters:
        - Target bit width: {bit_width}-bit
        - Outlier percentage to preserve: {outlier_percentage}%
        - Task type: {task_type}

        Your delegation should:
        1. Clearly specify what TCQAG-2 needs to demonstrate
        2. Provide any necessary context about TCQ that TCQAG-2 should know
        3. Request specific visualizations or examples that would be helpful
        4. Be professional but conversational in tone
        """

        try:
            response = client.chat.completions.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "You are TCQAG-1, the primary AI agent for Task-Circuit Quantization demonstration."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.7,
                max_tokens=800
            )
            return response.choices[0].message.content
        except Exception as e:
            return f"Error generating delegation: {str(e)}\n\nFallback delegation: TCQAG-2, please demonstrate the Task-Circuit Quantization process with {bit_width}-bit quantization while preserving {outlier_percentage}% of weights as outliers. Focus on the {task_type} task and show how TCQ improves performance compared to baseline methods."


class TCQAG2:
    """
    Secondary TCQ Agent (TCQAG-2) that performs the actual TCQ demonstration
    """
    def __init__(self):
        self.name = "TCQAG-2"
        self.role = "Task-Circuit Quantization Executor"

    def perform_demonstration(self, bit_width, outlier_percentage, task_type):
        """
        Performs a demonstration of the TCQ process
        """
        # Simulate processing time
        time.sleep(2)

        prompt = f"""
        You are TCQAG-2, an AI agent specialized in demonstrating Task-Circuit Quantization (TCQ).

        Perform a detailed demonstration of TCQ with the following parameters:
        - Target bit width: {bit_width}-bit
        - Outlier percentage to preserve: {outlier_percentage}%
        - Task type: {task_type}

        Your demonstration should include:
        1. A step-by-step walkthrough of the TCQ process with specific mathematical details
        2. How the saliency metric is calculated and applied using the formula: TACQ(Wij) = |Wij| · ∂L/∂Wij · |Q(Wij) - Wij|
        3. A detailed comparison of performance before and after TCQ with specific numbers from the paper
        4. Multiple specific examples relevant to the {task_type} task, including actual model outputs
        5. Advantages of TCQ over other quantization methods (GPTQ, SqueezeLLM, SPQR, SliM-LLM)
        6. A discussion of how TCQ identifies task-specific circuits in the model

        Make your explanation technical but clear, and include specific numerical examples where appropriate.
        For the {task_type} task, show a concrete example of how TCQ preserves model capabilities that other methods lose.

        Reference specific findings from the paper, such as:
        - In 2-bit settings, TCQ achieves 49.19% accuracy on MMLU compared to 34.84% for SliM-LLM
        - In 2-bit settings, TCQ achieves 36.11% accuracy on GSM8k compared to 20.12% for SliM-LLM
        - In 2-bit settings, TCQ achieves 21.92% accuracy on Spider while other methods degrade to near-zero
        """

        try:
            response = client.chat.completions.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "You are TCQAG-2, an AI agent specialized in demonstrating Task-Circuit Quantization."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.7,
                max_tokens=1200
            )
            return response.choices[0].message.content
        except Exception as e:
            return f"Error generating demonstration: {str(e)}\n\nFallback demonstration: I'll demonstrate Task-Circuit Quantization with {bit_width}-bit precision while preserving {outlier_percentage}% of weights. For {task_type} tasks, TCQ first identifies important weights using a saliency metric combining gradient information and quantization effects. By preserving critical weights in 16-bit precision while quantizing others to {bit_width}-bit, we maintain model performance while significantly reducing memory requirements."

    def simulate_tcq(self, weights, bit_width, outlier_percentage):
        """
        Simulates the TCQ process on a set of weights
        """
        # Create a copy of weights to avoid modifying the original
        quantized_weights = np.copy(weights)

        # Calculate saliency scores (simplified simulation)
        # In real TCQ, this would use gradient information and quantization effects
        saliency_scores = np.abs(weights) * np.random.normal(1, 0.2, size=weights.shape)

        # Identify top outlier_percentage% of weights by saliency score
        threshold = np.percentile(saliency_scores, 100 - outlier_percentage)
        outlier_mask = saliency_scores >= threshold

        # Quantize non-outlier weights to 2^bit_width levels
        non_outliers = ~outlier_mask
        levels = 2**bit_width
        min_val, max_val = weights[non_outliers].min(), weights[non_outliers].max()
        step = (max_val - min_val) / (levels - 1)

        for i in range(levels):
            level_val = min_val + i * step
            mask = non_outliers & (weights >= (level_val - step/2)) & (weights <= (level_val + step/2))
            quantized_weights[mask] = level_val

        return quantized_weights, outlier_mask
