requirements.txt
streamlit>=1.22.0
numpy>=1.24.3
matplotlib>=3.7.1
python-dotenv>=1.0.0
openai>=1.0.0
